[IN-276] feat:  add routs for Tab
