+++ src/features/profile/settings/getSettingsTabs.tsx
@@ -2,9 +2,12 @@ import React from 'react'
 
 import { Chrome, CreditCard, Person, Settings } from '@/assets/icons'
 import { TabsType, Typography } from '@/common'
-import { GeneralInfoPage } from '@/features/profile'
+// import { GeneralInfoPage } from '@/features/profile'
 import { Devices } from '@/features/profile/settings/devices'
+import Management from '@/features/profile/settings/management/ui/management'
 import { LocaleType } from '@locales/ru'
+
+import { GeneralInfoPage } from './generalInfo'
 
 export const getSettingsTabs = (
   t: LocaleType,
@@ -13,7 +16,7 @@ export const getSettingsTabs = (
   return [
     {
       content: <GeneralInfoPage />,
-      id: 'tab1',
+      id: 'general',
       title: isMobile ? (
         <Settings className={'w-full'} />
       ) : (
@@ -22,7 +25,7 @@ export const getSettingsTabs = (
     },
     {
       content: <Devices />,
-      id: 'tab2',
+      id: 'devices',
       title: isMobile ? (
         <Chrome className={'w-full'} />
       ) : (
@@ -31,7 +34,7 @@ export const getSettingsTabs = (
     },
     {
       content: <Typography variant={'reg16'}>Account Management</Typography>,
-      id: 'tab3',
+      id: 'management',
       title: isMobile ? (
         <Person className={'w-full'} />
       ) : (
@@ -40,7 +43,7 @@ export const getSettingsTabs = (
     },
     {
       content: <Typography variant={'reg16'}>My payments</Typography>,
-      id: 'tab4',
+      id: 'payments',
       title: isMobile ? (
         <CreditCard className={'w-full'} />
       ) : (
